+++ tests/precision_known_value_conversion_failure_dynamic_clean.adb
@@ -1,0 +1,7 @@
+procedure Precision_Known_Value_Conversion_Failure_Dynamic_Clean
+  (S : String)
+is
+   N : Integer;
+begin
+   N := Integer'Value (S);
+end Precision_Known_Value_Conversion_Failure_Dynamic_Clean;