+++ tests/precision_known_value_conversion_failure_integer_clean.adb
@@ -1,0 +1,5 @@
+procedure Precision_Known_Value_Conversion_Failure_Integer_Clean is
+   N : Integer;
+begin
+   N := Integer'Value ("123");
+end Precision_Known_Value_Conversion_Failure_Integer_Clean;