+++ tests/precision_known_value_conversion_failure_integer_finding.adb
@@ -1,0 +1,5 @@
+procedure Precision_Known_Value_Conversion_Failure_Integer_Finding is
+   N : Integer;
+begin
+   N := Integer'Value ("N/A");
+end Precision_Known_Value_Conversion_Failure_Integer_Finding;